Add code for obtaining contest data from API and storing in the database
This code fetches contest data from the Codeforces API and filters it based on the competition date. Only contest data with a competition date after today's date is stored in the database.

Here is a breakdown of the code sections:
- Fetching contest data from the API
- Filtering contests based on the competition date
- Establishing a connection to the database
- Mapping contest data to database parameters
- Executing the insert query and committing the changes

This code ensures that only relevant contest data is stored in the database, keeping the database up to date with upcoming contests.